Refactoring Styles
