+++ src/themes/Board.styles.ts
@@ -1,0 +1,15 @@
+import { StyleSheet } from 'react-native';
+
+export const boardStyle = StyleSheet.create({
+    container: {
+      alignItems: 'center',
+      justifyContent: 'center',
+    },
+    board: {
+      backgroundColor: '#000',
+    },
+    row: {
+      justifyContent: 'space-around',
+      flexDirection: 'row',
+    },
+  });
